\name{X1}
\alias{X1}
\docType{data}
\title{Example data X1
}
\description{
The package includes the official results of the state elections in Salzburg of 2009 and 2013. X1 contains the results of the election of 2009, Y1 the results of the election of 2013. Load data(X1) and data(Y1); 'loesung(X1,Y1)' calculates the voter shifts between the parties of the two elections. The example data set X1 contains the results of the following parties: SPOE, OEVP, FPOE, GRUENE, BZOE, OTHERS, NONVOTERS. Each row contains the result of one of the 119 municipalities.
}
\usage{data(X1)
}
\source{
http://www.salzburg.gv.at/20003stat/wahlen/ltw/index.htm#dl.5
}
\keyword{datasets}
