\name{Y1}
\alias{Y1}
\docType{data}
\title{
Example Data Y1
}
\description{
Official results of the state election in Salzburg 2013. The example dataset Y1 contains the results of the following parties: SPOE, OEVP, FPOE, GRUENE, TEAM, OTHERS, NONVOTERS. Each row contains the result of one of the 119 municipalities.
}
\usage{data(Y1)
}

\source{
http://www.salzburg.gv.at/20003stat/wahlen/ltw/index.htm#dl.5
}
\keyword{datasets}
